refactored Recipient into communication file
