+++ contracts/andromeda_splitter/src/contract.rs
@@ -1,7 +1,9 @@
 use crate::state::SPLITTER;
 use andromeda_protocol::{
     communication::encode_binary,
-    communication::{parse_message, AndromedaMsg, AndromedaQuery, ExecuteMsg as AndrExecute},
+    communication::{
+        parse_message, AndromedaMsg, AndromedaQuery, ExecuteMsg as AndrExecute, Recipient,
+    },
     error::ContractError,
     modules::{
         address_list::{on_address_list_reply, AddressListModule, REPLY_ADDRESS_LIST},
@@ -12,9 +14,9 @@ use andromeda_protocol::{
     ownership::{execute_update_owner, is_contract_owner, query_contract_owner, CONTRACT_OWNER},
     require,
     splitter::{
-        validate_recipient_list, AddressPercent, ExecuteMsg, InstantiateMsg, QueryMsg, Splitter,
+        validate_recipient_list, AddressPercent, ExecuteMsg, GetSplitterConfigResponse,
+        InstantiateMsg, QueryMsg, Splitter,
     },
-    splitter::{GetSplitterConfigResponse, Recipient},
 };
 use cosmwasm_std::{
     attr, entry_point, BankMsg, Binary, Coin, CosmosMsg, Deps, DepsMut, Env, MessageInfo, Reply,
@@ -140,7 +142,8 @@ fn execute_send(deps: DepsMut, info: MessageInfo) -> Result<Response, ContractEr
         }
         // ADO receivers must use AndromedaMsg::Receive to execute their functionality
         // Others may just receive the funds
-        let msg = match &recipient_addr.recipient {
+        let msg = recipient_addr.recipient.generate_msg(&deps, vec_coin)?;
+        /*let msg = match &recipient_addr.recipient {
             Recipient::ADO(recip) => SubMsg::new(WasmMsg::Execute {
                 contract_addr: deps.api.addr_validate(&recip.addr)?.to_string(),
                 msg: encode_binary(&AndrExecute::AndrReceive(AndromedaMsg::Receive(
@@ -152,7 +155,7 @@ fn execute_send(deps: DepsMut, info: MessageInfo) -> Result<Response, ContractEr
                 to_address: addr.clone(),
                 amount: vec_coin,
             })),
-        };
+        };*/
 
         submsg.push(msg);
     }
