+++ packages/andromeda_protocol/src/splitter.rs
@@ -1,34 +1,9 @@
-use crate::communication::{AndromedaMsg, AndromedaQuery};
+use crate::communication::{AndromedaMsg, AndromedaQuery, Recipient};
 use crate::error::ContractError;
 use crate::{modules::address_list::AddressListModule, require};
-use cosmwasm_std::{Binary, Uint128};
+use cosmwasm_std::Uint128;
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
-
-// ADOs use a default Receive message for handling funds, this struct states that the recipient is an ADO and may attach the data field to the Receive message
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
-pub struct ADORecipient {
-    pub addr: String,
-    pub msg: Option<Binary>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
-#[serde(rename_all = "snake_case")]
-pub enum Recipient {
-    Addr(String),
-    ADO(ADORecipient),
-}
-
-impl Recipient {
-    /// Creates an Addr Recipient from the given string
-    pub fn from_string(addr: String) -> Recipient {
-        Recipient::Addr(addr)
-    }
-    /// Creates an ADO Recipient from the given string with an empty Data field
-    pub fn ado_from_string(addr: String) -> Recipient {
-        Recipient::ADO(ADORecipient { addr, msg: None })
-    }
-}
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
 pub struct AddressPercent {
